Выделил MyStemManager
