+++ TagCloud/TagCloud/TagCloudBuilder.cs
@@ -21,6 +21,7 @@ namespace TagCloud
             _builder.RegisterType<WordLayouter>().As<IWordLayouter>();
             _builder.RegisterType<TxtFileProvider>().As<IFileProvider>();
             _builder.RegisterType<SimplePalette>().As<IPalette>();
+            _builder.RegisterType<MyStemManager>().AsSelf();
             _builder.RegisterType<TagCloud>().AsSelf();
 
             return this;
